requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
firebase-admin
python-dotenv
bcrypt